test(dataobj): fix test compilation after rebasing onto main
The rebase pulled in two main API changes the branch's tests hadn't adapted to:
- dataset.RowReaderOptions.Prefetch was renamed to PrefetchAllOnOpen (#24041).
- streams.ShardBucket now returns uint32; shardBucketFilter fields stay uint64.

Adjust the affected tests accordingly.

+++ pkg/querier/dataobj_cache_test.go
@@ -127,7 +127,7 @@ func TestDataObjCache_StreamLabels_ShardBucketFilter(t *testing.T) {
 
 	t.Run("filter to one stream's bucket prunes the others", func(t *testing.T) {
 		fooBucket := streams.ShardBucket(foo)
-		got, filtered, err := oo.streamLabels(ctx, wantSet(ids...), readQuery{shardBucket: &shardBucketFilter{from: fooBucket, to: fooBucket}})
+		got, filtered, err := oo.streamLabels(ctx, wantSet(ids...), readQuery{shardBucket: &shardBucketFilter{from: uint64(fooBucket), to: uint64(fooBucket)}})
 		require.NoError(t, err)
 		require.True(t, filtered)
 		// Exactly the streams whose bucket equals foo's survive.
